+++ src/server/game/Mails/Mail.h
@@ -131,7 +131,7 @@ class MailDraft
     public:                                                 // modifiers
         MailDraft& AddItem(Item* item);
         MailDraft& AddMoney(uint64 money) { m_money = money; return *this; }
-        MailDraft& AddCOD(uint32 COD) { m_COD = COD; return *this; }
+        MailDraft& AddCOD(uint64 COD) { m_COD = COD; return *this; }
 
     public:                                                 // finishers
         void SendReturnToSender(uint32 sender_acc, uint32 sender_guid, uint32 receiver_guid, SQLTransaction& trans);
